+++ lte/phy/lib/utils/src/bit.c
@@ -75,7 +75,7 @@ unsigned int bit_diff(char *x, char *y, int nbits) {
 }
 
 // Counts the number of ones in a word. K&R book exercise 2.9
-int bit_count(unsigned int n) {
+uint8_t bit_count(uint32_t n) {
   int c;
   for (c = 0; n; c++)
     n &= n - 1;
